example of adb
this example utilizes two separate tables to track the hose inventory and the hose testing; we then use a query with a left join to grab the latest test date, if one is available. 

Added two forms, one to display summary of hose inventory and last test date, and another to allow additions or edits to hose test history.